requirements.txt
flask==2.3.3
flask-cors==3.0.10
torch==2.7.1+cu118
torchvision==0.22.1+cu118
matplotlib==3.8.4
pytorch-msssim==0.2.1
scikit-learn==1.4.2
Pillow==11.3.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
scikit-image==0.22.0
albumentations==1.4.0
seaborn==0.13.2
fpdf==1.7.2
timm==1.0.19
